+++ src/AsmResolver.Workspaces.DotNet/Analyzers/Definition/MethodAnalyzer.cs
@@ -7,30 +7,12 @@ using AsmResolver.DotNet.Signatures;
 namespace AsmResolver.Workspaces.DotNet.Analyzers.Definition
 {
     /// <summary>
-    /// Analyzes a <see cref="MethodDefinition"/> for implicit base definitions, such as abstract methods in
-    /// base classes or methods in implemented interfaces.
+    /// Provides a default implementation for an <see cref="MethodDefinition"/> analyzer.
     /// </summary>
     public class MethodAnalyzer : ObjectAnalyzer<MethodDefinition>
     {
         /// <inheritdoc />
         protected override void Analyze(AnalysisContext context, MethodDefinition subject)
-        {
-            ScheduleMembersForAnalysis(context, subject);
-
-            if (!subject.IsVirtual)
-                return;
-
-            var index = context.Workspace.Index;
-            var node = index.GetOrCreateNode(subject);
-
-            foreach (var baseMethod in subject.FindBaseMethods(context.Workspace.Index))
-            {
-                var candidateNode = index.GetOrCreateNode(baseMethod);
-                node.ForwardRelations.Add(DotNetRelations.ImplementationMethod, candidateNode);
-            }
-        }
-
-        private static void ScheduleMembersForAnalysis(AnalysisContext context, MethodDefinition subject)
         {
             // Schedule parameters for analysis.
             if (context.HasAnalyzers(typeof(ParameterDefinition)))
@@ -50,7 +32,6 @@ namespace AsmResolver.Workspaces.DotNet.Analyzers.Definition
             {
                 context.ScheduleForAnalysis(subject.CilMethodBody);
             }
-
         }
     }
 }
